+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 * BREAKING CHANGE: The `m` parameter in in `Flay` and `Flayable` has been
   existentialized, to be chosen by the caller.
 
-* Added `Flayable1`, `flay1`, `trivial`, `collect1`, `Record`, `Unit`, `unit`,
- `zip1`, `zip`, `GFlay1`.
+* Added `Flayable1`, `trivial1`, `collect1`, `Terminal`, `zip1`, `zip`, `GFlay1`.
 
 
 # Version 0.1